+++ src/event/driver.rs
@@ -1,4 +1,4 @@
-use super::info::Event;
+use super::Event;
 
 ////////////////////////////////////////////////////////////////////////////////
