+++ src/event/info.rs
@@ -1,21 +1,26 @@
-use std::hash::Hash;
+use std::{hash::Hash, time::Duration};
 
-use crate::simulation::proc::{Address, ProcessHandle};
+use crate::simulation::proc::ProcessHandle;
 
-use super::time::TimeSegment;
+////////////////////////////////////////////////////////////////////////////////
+
+#[derive(Clone, Hash)]
+pub enum EventInfo {
+    UdpMessage(UdpMessage),
+    Timer(Timer),
+}
 
 ////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Clone)]
-pub struct UdpMessageInfo {
+pub struct UdpMessage {
     pub udp_msg_id: usize,
     pub from: ProcessHandle,
     pub to: ProcessHandle,
     pub content: String,
-    pub can_be_dropped: bool,
 }
 
-impl Hash for UdpMessageInfo {
+impl Hash for UdpMessage {
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
         self.from.address().hash(state);
         self.to.address().hash(state);
@@ -26,44 +31,15 @@ impl Hash for UdpMessageInfo {
 ////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Clone)]
-pub struct TimerInfo {
+pub struct Timer {
     pub timer_id: usize,
     pub with_sleep: bool,
     pub proc: ProcessHandle,
+    pub duration: Duration,
 }
 
-impl Hash for TimerInfo {
+impl Hash for Timer {
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.proc.address().hash(state);
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-#[derive(Clone, Hash)]
-pub enum EventInfo {
-    UdpMessageInfo(UdpMessageInfo),
-    TimerInfo(TimerInfo),
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-#[derive(Clone)]
-pub struct Event {
-    pub id: usize,
-    pub time: TimeSegment,
-    pub info: EventInfo,
-}
-
-impl Event {
-    pub fn new(id: usize, time: TimeSegment, info: EventInfo) -> Self {
-        Self { id, time, info }
-    }
-}
-
-impl Hash for Event {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.time.hash(state);
-        self.info.hash(state);
+        self.duration.hash(state);
     }
 }
