+++ src/event/outcome.rs
