--- src/event1/driver.rs
@@ -1,7 +1,0 @@
-use super::Event;
-
-////////////////////////////////////////////////////////////////////////////////
-
-pub trait EventDriver {
-    fn register_event(&mut self, event: &Event);
-}
