+++ src/simulation/log.rs
@@ -2,7 +2,7 @@ use std::{fmt::Display, time::Duration};
 
 use colored::Colorize;
 
-use crate::event1::time::TimeSegment;
+use crate::event::time::TimeSegment;
 
 use super::proc::Address;
